moved packages into modules, where practical
